Multiple fixes.
Fix demo on Python3 including when python=python3.  Use venv instead of
virtualenv if available.

Also create the venv in the demo folder.

Fix python -mplop.collector /path/to/script.py; this used to try to save
the profile output to profiles/path/to/script.py without creating the
intermediate folders (and thus would fail).  Now the profile is saved
next to the script if given by absolute path.